+++ packages/atomic/cypress/e2e/result-list/result-components/quickview.cypress.ts
@@ -142,6 +142,25 @@ describe('Quickview Component', () => {
     });
   });
 
+  describe('when used on pdf file inside a result list with exact phrase match', () => {
+    beforeEach(() => {
+      new TestFixture()
+        .with(addFacet({field: 'author'}))
+        .with(addSearchBox())
+        .withHash('q="the%20bank"&f-author=Jules%20Verne')
+        .with(addQuickviewInResultList())
+        .init();
+    });
+
+    it('should display only highlight button for the phrase with multiple occurrences', () => {
+      openModal();
+
+      QuickviewModalSelectors.keywordsSidebar()
+        .should('exist')
+        .should('contain.text', 'Navigate between 23 occurrences of the bank');
+    });
+  });
+
   describe('when used on an excel file with a query that returns invalid HTML', () => {
     beforeEach(() => {
       new TestFixture()
@@ -156,8 +175,7 @@ describe('Quickview Component', () => {
       openModal();
       QuickviewModalSelectors.keywordsSidebar()
         .should('exist')
-        .should('contain.text', 'Navigate between 6 occurrences of 1')
-        .should('contain.text', 'Navigate between 6 occurrences of 12');
+        .should('contain.text', 'Navigate between 12 occurrences of 1 12');
     });
   });
 });