+++ packages/headless/src/controllers/core/did-you-mean/headless-core-did-you-mean.test.ts
@@ -1,8 +1,9 @@
-import {configuration, didYouMean} from '../../../app/reducers';
+import {configuration} from '../../../app/common-reducers';
 import {
   applyDidYouMeanCorrection,
   enableDidYouMean,
 } from '../../../features/did-you-mean/did-you-mean-actions';
+import {didYouMeanReducer as didYouMean} from '../../../features/did-you-mean/did-you-mean-slice';
 import {
   buildMockSearchAppEngine,
   MockSearchEngine,