+++ packages/headless/src/controllers/recommendation/result-list/headless-recommendation-interactive-result.test.ts
@@ -1,5 +1,5 @@
 import {Result} from '../../../api/search/search/result';
-import {configuration} from '../../../app/reducers';
+import {configuration} from '../../../app/common-reducers';
 import {logRecommendationOpen} from '../../../features/recommendation/recommendation-analytics-actions';
 import {buildMockResult} from '../../../test';
 import {