+++ packages/headless/src/controllers/smart-snippet-questions-list/headless-smart-snippet-interactive-questions.ts
@@ -1,12 +1,13 @@
 import {Result} from '../../api/search/search/result';
-import {questionAnswering, search} from '../../app/reducers';
 import {SearchEngine} from '../../app/search-engine/search-engine';
 import {logOpenSmartSnippetSuggestionSource} from '../../features/question-answering/question-answering-analytics-actions';
 import {
   answerSourceSelector,
   relatedQuestionSelector,
 } from '../../features/question-answering/question-answering-selectors';
+import {questionAnsweringReducer as questionAnswering} from '../../features/question-answering/question-answering-slice';
 import {pushRecentResult} from '../../features/recent-results/recent-results-actions';
+import {searchReducer as search} from '../../features/search/search-slice';
 import {QuestionAnsweringSection} from '../../state/state-sections';
 import {loadReducerError} from '../../utils/errors';
 import {